requirements.txt
streamlit
transformers
torch
nltk
PyPDF2
python-docx
sentence-transformers
faiss-cpu
langdetect